Move declarations of non-static functions in cvgetinfo.c to own file
