+++ fontforgeexe/charinfo.c
@@ -27,6 +27,7 @@
  */
 
 #include "fontforgeui.h"
+#include "cvgetinfo.h"
 #include <ustring.h>
 #include <math.h>
 #include <utype.h>
