+++ fontforgeexe/cvpalettes.c
@@ -28,6 +28,7 @@
 #include "bitmapview.h"
 #include "collabclientui.h"
 #include "cvdebug.h"
+#include "cvgetinfo.h"
 
 int palettes_docked=1;
 int rectelipse=0, polystar=0, regular_star=1;
